requirements.txt
Django==1.8.3
psycopg2==2.6.1
pytz==2015.4
Pillow==3.0.0
django-imagekit==3.3
boto==2.38.0
https://bitbucket.org/david/django-storages/get/f153a70ba254.zip#egg=storages
djangorestframework==3.1.3
django-filter==0.10.0
markdown==2.6.2
django-push-notifications==1.3.1
requests==2.7.0
plivo==0.10.1
django-debug-toolbar==1.3.2
git+https://github.com/daviddrysdale/python-phonenumbers.git
